WIP: support when first inserted value can be formatted as decimal
